requirements.txt
requests==2.26.0
schedule==1.1.0
python-telegram-bot==13.8.1
environs==9.3.5
Django==3.2.9
celery[redis]==5.2.1
django-celery-beat==2.2.1
psycopg2-binary==2.9.2
black==22.3.0
Pillow==9.1.1
matplotlib==3.5.2